refactor!: remove peersyncing
Peersyncing was implemented but never used. Waku's SDS will replace it,
so it's being removed to reduce complexity.

iterates: #7016
