+++ messaging/types/handle_message_response.go
@@ -1,17 +1,6 @@
 package types
 
-import "crypto/ecdsa"
-
 type HandleMessageResponse struct {
-	Hash             []byte
-	StatusMessages   []*Message
-	DatasyncSender   *ecdsa.PublicKey
-	DatasyncAcks     [][]byte
-	DatasyncOffers   []DatasyncOffer
-	DatasyncRequests [][]byte
-}
-
-type DatasyncOffer struct {
-	GroupID   []byte
-	MessageID []byte
+	StatusMessages []*Message
+	DatasyncAcks   [][]byte
 }
