+++ multiaccounts/settings/database.go
@@ -409,7 +409,7 @@ func (db *Database) GetSettings() (Settings, error) {
 		test_networks_enabled, mutual_contact_enabled, profile_migration_needed, wallet_token_preferences_group_by_community, url_unfurling_mode,
 		mnemonic_was_not_shown, wallet_show_community_asset_when_sending_tokens, wallet_display_assets_below_balance,
 		wallet_display_assets_below_balance_threshold, wallet_collectible_preferences_group_by_collection, wallet_collectible_preferences_group_by_community,
-		peer_syncing_enabled, auto_refresh_tokens_enabled, last_tokens_update, news_feed_enabled, news_feed_last_fetched_timestamp, news_rss_enabled, backup_path,
+		auto_refresh_tokens_enabled, last_tokens_update, news_feed_enabled, news_feed_last_fetched_timestamp, news_rss_enabled, backup_path,
 		thirdparty_services_enabled, messages_backup_enabled
 	FROM
 		settings
@@ -489,7 +489,6 @@ func (db *Database) GetSettings() (Settings, error) {
 		&s.DisplayAssetsBelowBalanceThreshold,
 		&s.CollectibleGroupByCollection,
 		&s.CollectibleGroupByCommunity,
-		&s.PeerSyncingEnabled,
 		&s.AutoRefreshTokensEnabled,
 		&lastTokensUpdate,
 		&s.NewsFeedEnabled,
@@ -758,24 +757,12 @@ func (db *Database) GetTestNetworksEnabled() (result bool, err error) {
 	return result, err
 }
 
-func (db *Database) SetPeerSyncingEnabled(value bool) error {
-	return db.SaveSettingField(PeerSyncingEnabled, value)
-}
-
 func (db *Database) SetSyncingOnMobileNetwork(value bool) error {
 	err := db.SaveSettingField(SyncingOnMobileNetwork, value)
 	if err != nil {
 		return err
 	}
 	return db.SaveSettingField(RememberSyncingChoice, true)
-}
-
-func (db *Database) GetPeerSyncingEnabled() (result bool, err error) {
-	err = db.makeSelectRow(PeerSyncingEnabled).Scan(&result)
-	if err == sql.ErrNoRows {
-		return result, nil
-	}
-	return result, err
 }
 
 func (db *Database) GetTokenGroupByCommunity() (result bool, err error) {
