+++ protocol/common/feature_flags.go
@@ -15,15 +15,6 @@ type FeatureFlags struct {
 	// DisableAutoMessageLoop disables auto message loop
 	DisableAutoMessageLoop bool
 
-	// ResendRawMessagesDisabled indicates whether we should be disabling sending raw messages
-	ResendRawMessagesDisabled bool
-
-	// StoreNodesDisabled indicates whether we should fetch messages from store nodes
-	StoreNodesDisabled bool
-
-	// Peersyncing indicates whether we should advertise and sync messages with other peers
-	Peersyncing bool
-
 	// EnableMercuryoProvider indicates whether we should enable the Mercuryo provider in the Wallet
 	EnableMercuryoProvider bool
