--- protocol/requests/toggle_peer_syncing.go
@@ -1,9 +1,0 @@
-package requests
-
-type TogglePeerSyncingRequest struct {
-	Enabled bool `json:"enabled"`
-}
-
-func (a *TogglePeerSyncingRequest) Validate() error {
-	return nil
-}
